Replace raw "color escape" unicode character with escaped variant
Fixes display of the broken messages and ensures that the project
files contain valid characters for their defined character encoding.

Could have wrapped it in `ChatColor.translateAlternateColorCodes` or
have used the constant from that same class but there are simply too
many references.

This seems like a decent enough fix taking the size of the module
in account.